+++ packages/python_ffi_cpython_dart/lib/src/cache/download_entry.dart
@@ -1,0 +1,10 @@
+part of python_ffi_cpython_dart;
+
+final class _DownloadEntry {
+  _DownloadEntry({required this.url, required this.sha256});
+
+  final String url;
+  final String sha256;
+
+  String get filename => url.split("/").last;
+}