+++ packages/python_ffi_cpython_dart/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Regenerated Python C-API bindings with all Python header files.
 - Made Python runtime initialization more robust.
+- The Python runtime is now downloaded and cached if the `libPyth` parameter is not specified.
+- The Python stdlib is now downloaded and cached.
 
 ## 0.3.0+1